FIXED INT ISSUE, MADE INTS FLOATS

+++ F_forcecalibration.ino
@@ -12,15 +12,15 @@ void loop() {
             total = total + analogRead(0);
             delay(100);
         }
-        int averageForce = total/10;
+        float averageForce = total/10;
         Serial.println("Average raw force value: " + String(averageForce)) + ".";
-        int massInGrams = ((106)*averageForce)/(1023-averageForce);
+        float massInGrams = ((106)*averageForce)/(1023-averageForce);
         Serial.println("Average mass in grams: " + String(massInGrams)) + ".";
-        int massInKilos = massInGrams / 1000;
+        float massInKilos = massInGrams / 1000;
         Serial.println("Average mass in kilos: " + String(massInKilos)) + ".";
-        int weightInNewtons = massInGrams / 101;
+        float weightInNewtons = massInGrams / 101;
         Serial.println("Average weight in newtons: " + String(weightInNewtons)) + ".";
-        int weightInPounds = weightInNewtons / 4.4;
+        float weightInPounds = weightInNewtons / 4.4;
         Serial.println("Average weight in pounds: " + String(weightInPounds)) + ".";
     }
     delay(200);
